Add files via upload
Here is the linear regression on homicides. You'll notice that the regression line doesn't seem to take into account the historical spike in 2016 and the steady decreases afterward. I think this may be a useful opportunity to talk about the limits of data in some cases. Tomorrow I'll add a regression analysis just on 2016 onward.